Project/Server/config.py
import os

from dotenv import load_dotenv

load_dotenv()

class Config:
    MONGO_URI = os.getenv("MONGO_URI", "mongodb://localhost:27017/socialnotes")
    JWT_SECRET_KEY = os.getenv("JWT_SECRET_KEY", "change-me-to-a-secure-key")
    JWT_ACCESS_TOKEN_EXPIRES = os.getenv("JWT_ACCESS_TOKEN_EXPIRES", 3600)

