Project/Server/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
Flask-JWT-Extended==4.4.4
pymongo==4.4.0
python-dotenv==1.0.0

